apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
pydantic==2.9.0
PyJWT==2.9.0
cryptography==43.0.0
